streamlit_app/pyproject.toml
[project]
name = "insurance-analytics-dashboard"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "snowflake-connector-python>=3.3.0",
    "streamlit>=1.41.0",
    "snowflake-ml-python>=1.8.0",
    "altair>=5.0.0",
    "pandas>=2.0.0",
]
